Feature/rework interfaces (#7)
* ref(create chunks) Add parallel workers

* opti(int vector) Optimize GC and memory alloc

* fix(create chunk) race condition

* fix(vector) repair incorrect insert

* ref(memusage) add field to enable memusage analysis

* ref(test) better test names

* add(makefile) add command to test race conditions

* add(ignore) untrack benchmark files

* ref(merge sort) export results to file
ref(merge sort) minimize chunks min value
ref(memusage) become an option

* ref(create chunks) Add channel with batches to sort in parallel

* opti(main) find good parameter (speed:1min10, ram:80M)

* ref(vector) Remove unecessary sort function

* fix(test) Repair batching channel test
ref(vector) Remove useless methods

* ref(vector) Replace all empty interfaces with concrete types
new(tsv model) Allow merge sort on tsv files

* ref(allocate) Beautify slice vector setup

* Increase output buffer size

* ref (tsv vector) Remove unusued struct
fix(create chunks) Fix wrong context

* Add docker setup
Remove deprectaed function

+++ file/utils.go
@@ -1,30 +1,32 @@
 package file
 
 import (
-	"io/ioutil"
 	"os"
 	"path"
+	"strings"
 
 	"github.com/pkg/errors"
 )
 
-// clearFolder Remove all files from a folder.
-func clearFolder(folder string) error {
+// clearChunkFolder Remove all files from a folder.
+func clearChunkFolder(folder string) error {
 	fn := "clear folder"
 	err := os.MkdirAll(folder, os.ModePerm)
 	if err != nil {
 		return errors.Wrap(err, fn)
 	}
-	dir, err := ioutil.ReadDir(folder)
+	dir, err := os.ReadDir(folder)
 	if err != nil {
 		return errors.Wrap(err, fn)
 	}
 	for _, d := range dir {
+		if !strings.HasPrefix(d.Name(), "chunk") {
+			continue
+		}
 		err = os.RemoveAll(path.Join(folder, d.Name()))
 		if err != nil {
 			return errors.Wrap(err, fn)
 		}
 	}
-
 	return nil
 }
